+++ backend/chidic-domain/src/main/kotlin/com/chidicdomain/domain/service/follow/FollowService.kt
@@ -1,9 +1,9 @@
 package com.chidicdomain.domain.service.follow
 
-import com.chidiccommon.dto.FollowCountResponse
+import com.chidicdomain.dto.FollowCountDto
 
 interface FollowService {
-    fun getFollowerAndFolloweeCount(userId: Long): FollowCountResponse
+    fun getFollowerAndFolloweeCount(userId: Long): FollowCountDto
     fun follow(userId: Long, followingId: Long)
     fun unfollow(userId: Long, followeeId: Long)
 }
