+++ backend/chidic-domain/src/main/kotlin/com/chidicdomain/dto/FollowCountDto.kt
@@ -1,0 +1,6 @@
+package com.chidicdomain.dto
+
+data class FollowCountDto(
+    val followerCount: Long,
+    val followingCount: Long
+)
